Moves ToiConf out to its own files
